add vortex and gravity
