+++ src/camera/camera.cpp
