+++ src/utils.hpp
@@ -13,5 +13,7 @@ enum dimension_enum {
 enum right_click_action_enum {
 	SPAWN_PARTICLES,
 	REMOVE_PARTICLES,
-	ADD_FORCE
+	ADD_RADIAL_FORCE,
+	ADD_VORTEX_FORCE,
+	ADD_GRAVITY_FORCE
 };
